+++ src/app/features/cart/components/cart-item/cart-item.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { CartItem } from '../../models/cart-item';
 
 @Component({