+++ src/app/shared/shared.module.ts
@@ -1,12 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { HighlightDirective } from './directives/highlight.directive';
 
-
+const EXPORTED_COMPONENTS = [
+  HighlightDirective,
+];
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    ...EXPORTED_COMPONENTS,
+  ],
+  exports: [
+    ...EXPORTED_COMPONENTS,
+  ],
   imports: [
-    CommonModule
+    CommonModule,
   ]
 })
 export class SharedModule { }