+++ src/app/shp.component.ts
@@ -1,10 +1,17 @@
-import { Component } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
 
 @Component({
   selector: 'shp-root',
   templateUrl: './shp.component.html',
   styleUrls: ['./shp.component.scss']
 })
-export class ShpComponent {
-  title = 'shop';
+export class ShpComponent implements AfterViewInit {
+  title = 'My Best Shop Market';
+
+  @ViewChild('appTitle')
+  titleElement!: ElementRef<HTMLHeadingElement>;
+
+  ngAfterViewInit(): void {
+    this.titleElement.nativeElement.innerHTML = this.title;
+  }
 }